Restore L profile definition
Keep testing XL profile.
