+++ src/test/kotlin/com/atlassian/performance/tools/hardware/HardwareExplorationIT.kt
@@ -1,11 +1,8 @@
 package com.atlassian.performance.tools.hardware
 
-import com.amazonaws.regions.Regions
 import com.amazonaws.services.ec2.model.InstanceType.*
 import com.amazonaws.services.s3.transfer.TransferManagerBuilder
 import com.atlassian.performance.tools.aws.api.Investment
-import com.atlassian.performance.tools.aws.api.StorageLocation
-import com.atlassian.performance.tools.awsinfrastructure.S3DatasetPackage
 import com.atlassian.performance.tools.hardware.IntegrationTestRuntime.aws
 import com.atlassian.performance.tools.hardware.IntegrationTestRuntime.logContext
 import com.atlassian.performance.tools.hardware.IntegrationTestRuntime.taskName
@@ -14,23 +11,13 @@ import com.atlassian.performance.tools.hardware.failure.BugAwareTolerance
 import com.atlassian.performance.tools.hardware.guidance.DbExplorationGuidance
 import com.atlassian.performance.tools.hardware.guidance.ExplorationGuidance
 import com.atlassian.performance.tools.hardware.guidance.JiraExplorationGuidance
-import com.atlassian.performance.tools.infrastructure.api.database.PostgresDatabase
-import com.atlassian.performance.tools.infrastructure.api.dataset.Dataset
-import com.atlassian.performance.tools.infrastructure.api.jira.JiraHomePackage
 import com.atlassian.performance.tools.jvmtasks.api.TaskTimer.time
-import com.atlassian.performance.tools.lib.LicenseOverridingDatabase
-import com.atlassian.performance.tools.lib.overrideDatabase
 import com.atlassian.performance.tools.lib.s3cache.S3Cache
-import com.atlassian.performance.tools.lib.toExistingFile
 import com.atlassian.performance.tools.lib.workspace.GitRepo2
-import com.atlassian.performance.tools.virtualusers.api.TemporalRate
-import com.atlassian.performance.tools.virtualusers.api.VirtualUserLoad
 import org.apache.logging.log4j.Logger
 import org.eclipse.jgit.api.Git
 import org.junit.Test
 import java.io.File
-import java.net.URI
-import java.nio.file.Paths
 import java.time.Duration
 
 const val jiraAdminPassword = "MasterPassword18"
@@ -38,44 +25,6 @@ const val jiraAdminPassword = "MasterPassword18"
 class HardwareExplorationIT {
 
     private val logger: Logger = logContext.getLogger(this::class.java.canonicalName)
-
-    private val sevenMillionIssues = StorageLocation(
-        uri = URI("s3://jpt-custom-postgres-xl/dataset-7m"),
-        region = Regions.EU_WEST_1
-    ).let { location ->
-        Dataset(
-            label = "7M issues",
-            database = PostgresDatabase(
-                source = S3DatasetPackage(
-                    artifactName = "database.tar.bz2",
-                    location = location,
-                    unpackedPath = "database",
-                    downloadTimeout = Duration.ofMinutes(40)
-                ),
-                dbName = "atldb",
-                dbUser = "postgres",
-                dbPassword = "postgres"
-            ),
-            jiraHomeSource = JiraHomePackage(
-                S3DatasetPackage(
-                    artifactName = "jirahome.tar.bz2",
-                    location = location,
-                    unpackedPath = "jirahome",
-                    downloadTimeout = Duration.ofMinutes(40)
-                )
-            )
-        )
-    }.overrideDatabase { originalDataset ->
-        val localLicense = Paths.get("jira-license.txt")
-        LicenseOverridingDatabase(
-            originalDataset.database,
-            listOf(
-                localLicense
-                    .toExistingFile()
-                    ?.readText()
-                    ?: throw  Exception("Put a Jira license to ${localLicense.toAbsolutePath()}")
-            ))
-    }
 
     private val jiraInstanceTypes = listOf(
         C48xlarge
@@ -143,17 +92,7 @@ class HardwareExplorationIT {
     private fun explore(
         guidance: ExplorationGuidance
     ): List<HardwareExplorationResult> = HardwareExploration(
-        scale = ApplicationScale(
-            description = "Jira L profile",
-            dataset = sevenMillionIssues,
-            load = VirtualUserLoad.Builder()
-                .virtualUsers(150)
-                .ramp(Duration.ofSeconds(90))
-                .flat(Duration.ofMinutes(20))
-                .maxOverallLoad(TemporalRate(30.0, Duration.ofSeconds(1)))
-                .build(),
-            vuNodes = 12
-        ),
+        scale = JIRA_EXTRA_LARGE,
         guidance = guidance,
         maxApdexSpread = 0.10,
         maxErrorRate = 0.05,
